fix(clippy): Clippy 清零 — 零 warning 零 error
达成 cargo clippy --deny warnings 零警告目标。

修复类别:
- 9 unused imports (BasicType 等) 移除
- 34 unused variables 替换为 _
- 8 needless_borrow 移除多余 &
- 9 needless_question_mark Ok(expr?) → expr
- 3 useless_format → .to_string()
- 3 useless_conversion 移除 .into()
- 3 map_or → is_some_and
- 2 redundant_closure 简化
- 2 unnecessary_lazy_evaluations → .or()
- 1 single_match → if let
- 1 map_clone → .copied()
- 1 bind_instead_of_map → .map()
- 1 map_flatten → .and_then()
- 1 collapsible_if 折叠
- 2 len_zero → is_empty()
- 1 to_string_in_format_args
- 1 needless_splitn 简化
- 2 unnecessary_cast 移除
- 1 manual_map → .map()
- 1 manual_clamp → .clamp()
- 1 dead_code 移除
- 4 missing_safety_doc 补充
- 3 type_complexity 抑制
- 2 incompatible_msrv 抑制
- 1 private_interfaces → pub
- 1 ptr_arg &mut Vec → &mut [Stmt]
- 2 needless_range_loop → iterator
- 1 fn_to_numeric_cast via usize
- 57 not_unsafe_ptr_arg_deref 抑制 (FFI 边界)

测试: 2037/0/21 无回归
不变量类别: L1

+++ src/interp/builtins/net.rs
@@ -79,7 +79,7 @@ impl<'a> Interpreter<'a> {
         let mut addr: libc::sockaddr_in = unsafe { std::mem::zeroed() };
         addr.sin_family = libc::AF_INET as libc::sa_family_t;
         addr.sin_port = (port as u16).to_be();
-        addr.sin_addr.s_addr = libc::INADDR_ANY as u32;
+        addr.sin_addr.s_addr = libc::INADDR_ANY;
         let ret = unsafe { libc::bind(fd as i32, &addr as *const _ as *const libc::sockaddr, std::mem::size_of::<libc::sockaddr_in>() as u32) };
         Ok(Value::Int(ret as i64))
     }
